falabellabackend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
Faker==37.5.3
inflection==0.5.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
